remove unity logic from container and put it back in attachedContainer
